assessment-service/app/requirements.txt
fastapi
uvicorn[standard]
httpx
psycopg2-binary
pydantic
python-multipart
Pillow
